chore: Drop FFmpeg capabilities
This change is being made because processing recorded files with FFmpeg
in this state does not fit the current program structure. With a lot of followed models, the script would acquire huge processing power. Moreover, test shows that not every record needs re-encoding, invaluable records are mostly discarded. The idea will be reconsidered
at a later time.